♻️ Move Pf2eAttack into Pf2eJsonNodeReader
Also add a transformListFrom method to JsonNodeReader which returns
a list of strings rather than a joined string.
