+++ src/main/java/dev/ebullient/convert/tools/pf2e/Json2QuteHazard.java
@@ -34,7 +34,7 @@ public class Json2QuteHazard extends Json2QuteBase {
                 Pf2eHazard.routine.transformTextFrom(rootNode, "\n", index),
                 Pf2eHazard.defenses.getDefensesFrom(rootNode, this),
                 Pf2eHazard.attacks.streamFrom(rootNode)
-                        .map(n -> Pf2eAttack.createInlineAttack(n, this))
+                        .map(n -> Pf2eJsonNodeReader.Pf2eAttack.createAttack(n, this))
                         .toList(),
                 Pf2eHazard.abilities.streamFrom(rootNode)
                         .map(n -> Pf2eAbility.createEmbeddedAbility(n, this))
