+++ src/main/java/dev/ebullient/convert/tools/pf2e/Pf2eTypeReader.java
@@ -6,8 +6,6 @@ import static dev.ebullient.convert.StringUtil.toTitleCase;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -18,7 +16,6 @@ import dev.ebullient.convert.qute.NamedText;
 import dev.ebullient.convert.tools.JsonNodeReader;
 import dev.ebullient.convert.tools.Tags;
 import dev.ebullient.convert.tools.pf2e.qute.QuteDataSkillBonus;
-import dev.ebullient.convert.tools.pf2e.qute.QuteInlineAttack;
 import dev.ebullient.convert.tools.pf2e.qute.QuteItem.QuteItemWeaponData;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 
@@ -303,81 +300,6 @@ public interface Pf2eTypeReader extends JsonSource {
                     convert.streamPropsExcluding(source, std, note)
                             .collect(Collectors.toMap(e -> convert.replaceText(e.getKey()), e -> e.getValue().asInt())),
                     note.getTextFrom(source).map(convert::replaceText).map(List::of).orElse(List.of()));
-        }
-    }
-
-    /**
-     * Example JSON input for a creature:
-     *
-     * <pre>
-     *     "range": "Melee",
-     *     "name": "jaws",
-     *     "attack": 32,
-     *     "traits": ["evil", "magical", "reach 10 feet"],
-     *     "effects": ["essence drain", "Grab"],
-     *     "damage": "3d8+9 piercing plus 1d6 evil, essence drain, and Grab",
-     *     "types": ["evil", "piercing"]
-     * </pre>
-     *
-     * An example for a hazard with a complicated effect:
-     *
-     * <pre>
-     *     "type": "attack",
-     *     "range": "Ranged",
-     *     "name": "eye beam",
-     *     "attack": 20,
-     *     "traits": ["diving", "evocation", "range 120 feet"],
-     *     "effects": [
-     *         "The target is subjected to one of the effects summarized below.",
-     *         {
-     *             "type": "list",
-     *             "items": [{
-     *                 "type": "item",
-     *                 "name": "Green Eye Beam",
-     *                 "entries": ["(poison) 6d6 poison damage (DC24 basic Reflex save)"],
-     *             }, ...],
-     *         },
-     *     ],
-     *     "types": ["electricity", "fire", "poison", "acid"]
-     * </pre>
-     *
-     */
-    enum Pf2eAttack implements Pf2eJsonNodeReader {
-        name,
-        attack,
-        activity,
-        damage,
-        effects,
-        range,
-        types,
-        noMAP;
-
-        public static QuteInlineAttack createInlineAttack(JsonNode node, JsonSource convert) {
-            List<String> effects = new ArrayList<>();
-            convert.appendToText(effects, Pf2eAttack.effects.getFrom(node), null);
-
-            // Either the effects are a list of short descriptors which are also included in the damage, or they are a
-            // long multi-line description of a complicated effect.
-            String formattedDamage = damage.replaceTextFrom(node, convert);
-            String multilineEffect = null;
-            if (effects.stream().anyMatch(Predicate.not(formattedDamage::contains))) {
-                multilineEffect = String.join("\n", effects); // Preserve empty strings for line breaks
-                effects = List.of();
-            }
-
-            return new QuteInlineAttack(
-                    name.replaceTextFrom(node, convert),
-                    Optional.ofNullable(activity.getActivityFrom(node, convert))
-                            .orElse(Pf2eActivity.single.toQuteActivity(convert, "")),
-                    QuteInlineAttack.AttackRangeType.valueOf(range.getTextOrDefault(node, "Melee").toUpperCase()),
-                    attack.getIntFrom(node).orElse(null),
-                    formattedDamage,
-                    types.replaceTextFromList(node, convert),
-                    convert.collectTraitsFrom(node, null),
-                    effects,
-                    multilineEffect,
-                    noMAP.booleanOrDefault(node, false) ? List.of() : List.of("no multiple attack penalty"),
-                    convert);
         }
     }
